Address PR review comments
Round 2 copilot feedback on the NAT resilience suite:

- Extract shared NAT test helpers (trackConsole, initPage, rebindTracker,
  waitForPeerConnection, waitForMesh, PageHandle) into a new
  e2e/integration/helpers/nat-helpers.ts module so nat-traversal.spec.ts
  and nat-resilience.spec.ts can't drift on tracker semantics or init
  timeouts.
- Bump the PEER_ID: wait inside initPage from 5s to 30s
  (PEER_ID_WAIT_MS), and surface the value via an option so callers can
  override it. The previous 5s timeout was much stricter than the
  baseline nat-traversal suite and could flake on slower machines or
  after container restarts.
- Tighten the RUN_NAT_RESTART env-var opt-in to compare strictly against
  '1' instead of !!process.env.RUN_NAT_RESTART, so values like "0" don't
  accidentally enable the slow/flaky relay-restart scenario in CI.
